# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bridge-Watch API",
    "description": "Real-time Stellar bridge monitoring API. Provides bridge statuses, asset health, liquidity, price feeds, alert management, analytics, and administrative controls.\n\n## Authentication\nProtected endpoints require an API key in the `x-api-key` header.\n\n## Rate Limiting\nAll endpoints are rate-limited per IP using a sliding-window algorithm backed by Redis. Exceeding the limit returns **429 Too Many Requests** with a `Retry-After` header.\n\n## Versioning\nAll REST endpoints are prefixed with `/api/v1/`. Breaking changes increment the version segment.\n\n## Error Format\n```json\n{ \"error\": \"Short label\", \"message\": \"Human-readable description\" }\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Bridge-Watch Team",
      "url": "https://github.com/StellaBridge/Bridge-Watch"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" },
    { "url": "https://api.bridge-watch.io", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for protected endpoints"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/watchlists/{userId}": {
      "get": {
        "tags": ["Watchlists"],
        "summary": "Get all watchlists for a user",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Watchlists" } }
      },
      "post": {
        "tags": ["Watchlists"],
        "summary": "Create a watchlist",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "isDefault": { "type": "boolean" } } } } } },
        "responses": { "200": { "description": "Created watchlist" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/watchlists/{userId}/{id}": {
      "delete": {
        "tags": ["Watchlists"],
        "summary": "Delete a watchlist",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      },
      "patch": {
        "tags": ["Watchlists"],
        "summary": "Update a watchlist",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Updated" }, "400": { "description": "Error" } }
      }
    },
    "/api/v1/preferences/{userId}": {
      "get": {
        "tags": ["Preferences"],
        "summary": "Get all preferences for a user",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Preferences" } }
      }
    },
    "/api/v1/preferences/{userId}/{category}/{key}": {
      "get": {
        "tags": ["Preferences"],
        "summary": "Get a single preference value",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "category", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Preference value" }, "400": { "description": "Invalid category" }, "404": { "description": "Key not found" } }
      },
      "put": {
        "tags": ["Preferences"],
        "summary": "Set a preference value",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "category", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["value"], "properties": { "value": {} } } } } },
        "responses": { "200": { "description": "Updated preferences" } }
      },
      "delete": {
        "tags": ["Preferences"],
        "summary": "Reset (delete) a preference key",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "category", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "key", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Reset preferences" } }
      }
    },
    "/api/v1/preferences/{userId}/bulk": {
      "patch": {
        "tags": ["Preferences"],
        "summary": "Bulk update preferences",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Updated preferences" }, "409": { "description": "Version conflict" } }
      }
    },
    "/api/v1/preferences/{userId}/export": {
      "get": {
        "tags": ["Preferences"],
        "summary": "Export user preferences",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Exported preferences" } }
      }
    },
    "/api/v1/preferences/{userId}/import": {
      "post": {
        "tags": ["Preferences"],
        "summary": "Import user preferences",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Imported preferences" } }
      }
    },
    "/api/v1/preferences/{userId}/stream": {
      "get": {
        "tags": ["Preferences"],
        "summary": "Stream preference change events (SSE)",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "SSE stream", "content": { "text/event-stream": {} } } }
      }
    },
...
  }
}